refactor: removed GoToDateModal and replaced it with form component to slot into dialog box
